Add a cache and search the cache first for movies
Search the cache for an existing key that contains the search string.
Then select the longest matching key, and search within that subset.
This works because the search query in the backend is a string contains.
A search string that expands on a previous search string will always return
a subset of the previous results (e.g. the results for "Alien" will always
be a subset of the results for "A").